+++ Sources/Tracking/CustomerIO.swift
@@ -4,6 +4,9 @@ import Foundation
 public protocol CustomerIOInstance: AutoMockable {
     var siteId: String? { get }
 
+    /// Get the current configuration options set for the SDK.
+    var config: SdkConfig? { get }
+
     func identify(
         identifier: String,
         body: [String: Any]
@@ -89,6 +92,11 @@ public extension CustomerIOInstance {
  You must call `CustomerIO.initialize` to use the features of the SDK.
  */
 public class CustomerIO: CustomerIOInstance {
+    /// The current version of the Customer.io SDK.
+    public static var version: String {
+        SdkVersion.version
+    }
+
     public var siteId: String? {
         diGraph?.siteId
     }
@@ -236,6 +244,10 @@ public class CustomerIO: CustomerIOInstance {
             .info(
                 "Customer.io SDK \(SdkVersion.version) initialized and ready to use for site id: \(siteId)"
             )
+    }
+
+    public var config: SdkConfig? {
+        implementation?.config
     }
 
     /**